reorg files + fix docs
