+++ .github/FUNDING.yml
@@ -1,2 +1,1 @@
-github: [tanner0101] # loganwright, joscdk
-open_collective: vapor
+github: [tanner0101]
