+++ Sources/JWTKit/HMAC/HMACError.swift
